Update to use scrypt and HKDF
Instead of using PBKDF2, the key derivation is now done with scrypt.  HKDF is also used to derive 64-byte chunks to comprise the 32-megabyte key based on this key instead of the convoluted loop utilizing rand() that was previously used.  HKDF is now also used to generate an auth key for use by the authentication and password checking.  Since rand() is no longer used, there's no need to generate a seed value and so the counter variable is initialized to 0 like is traditionally done, rather than deriving a random value to set it to with PBKDF2 and seed rand() with as was being done previously.