Made it so password enry isn't cleared on non-matching passwords
The password entry boxes are cleared as soon as encryption/decryption is done, but this was also happening if the passwords didn't match, which was kind of annoying since that means they'd have to be entered in completely over again.  Now if the passes don't match they're left in the entry box to be edited.