Changed cipher algorithm back to a simple counter
I modified the cipher algorithm to multiply the counter by the key value, but this caused problems if the key had any zero values because as soon as the counter was multiplied by 0, it would become 0, and then any key value multiplied by the counter from then on would also become 0. I tried to mitigate this by making sure no value was 0, but it ended up with an unevenly distributed number of byte values in the resultant cipher text anyway, so I revered to using a simple incrementing counter variable. I also added a nonce value back into the mix just to increase complexity a little bit, with the nonce being a random number derived from the key.