Added features and changes to GUI
Added the ability to use a "one-time-pad" derived from /dev/urandom or other files, for example a hardware based RNG. The output file will be accompanied by a file with the same name, but with the .pad extension in this case.

Otherwise, also added the ability for the key file or the one-time-pad to be accompanied by a password. If using a keyfile, the password will be used to generate a key, and this key will be prepended into the keystream that the keyfile generates. Otherwise, if a one-time-pad is selected, the password will only be used to generate unique counterInt and nonceInt values. The thinking here is that the one-time-pad should remain "pure" and not have any deterministic key derived from a password added to it, but considering that the counterInt and noneInt are also derived from a deterministic value, it might not really make that much difference

Also added these new features to the GUI, as well as the ability to specify the key size derived from a password, as well as the buffer sizes used for generating a MAC and for the the input and output of the file itself.