Fixed memory leak in yaxa() if using one-time-pad
The otpBuffer is passed as a pointer into yaxa(), but since it becomes its own pointer, it is lost in the calling doCrypt() function. Changed otpBuffer to be global instead. Could probably just make yaxa use a pointer to a pointer or **otPBuffer but don't want to since a global ariable is just simpler.