Fixed old values not being replaced after encrypt/decrypt button pressed in GUI
If a user were to encrypt a file with a non-default parameter, and then try to encrypt another file with a default parameter, the old non-default would still be used due to the way the conditions were taking values from the GUI entries. For example, if one selected a non-default key size for one file, and then selected the default 32 MB to encrypt another file, the non-default value would be used instead of the default 32 MB file.