Reduced redundant code in workThread in GUI
Before, the workThread() code tested for two conditionals, if action equaled 'e' or 'd' to determine if it was doing encryption or decryption. A large part of the code under each conditional was fairly redundant, so I changed it to have less blocks of redundant code and to do specific operations based on if it was encrypting or decrypting.