Added nested structures and fixed otpBuffer issue
Added nested structures to add mre context to various structure members.

Also made otpBuffer a global variable again because it was again causing incorrect decryptions between programs due to a memory error. Interestingly enough this memory error wasn't detected by valgrind or ASan, but I still think it is because it needs to be a pointer to a pointer. Will have to investigate it later.