Added otpBuffer back to be a buffer local to doCrypt
After extensive testing it seems that whatever memory issue was causing a encryption/decryption discrepancy between the GUi and the CLI if using a one-time-pad is no longer an issue. Now otpBuffer is decared and allocated in doCrypt and passed to yaxa directly rather than through a structure or global variable.